// lib/api/generated/models/apiKeyAppResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyResponse } from "./apiKeyResponse";
import type { EmptyAppResponse } from "./emptyAppResponse";

export type ApiKeyAppResponse = EmptyAppResponse & {
  data: ApiKeyResponse;
};