// lib/api/generated/models/apiKeyListAppResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Realillust API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyListResponse } from "./apiKeyListResponse";
import type { EmptyAppResponse } from "./emptyAppResponse";

export type ApiKeyListAppResponse = EmptyAppResponse & {
  data: ApiKeyListResponse;
};