// lib/api/generated/models/authUser.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Realillust API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  /** @nullable */
  email: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  image: string | null;
  onboardingCompleted: boolean;
}